+++ SIL.Core/IO/RobustIO.cs
@@ -4,7 +4,7 @@ using System.Drawing;
 using System.Drawing.Imaging;
 using System.IO;
 using System.Xml.Linq;
-using SIL.Retry;
+using SIL.Code;
 
 namespace SIL.IO
 {
